+++ task/executor/pkg/config_resolver.go
@@ -65,17 +65,18 @@ func (r *configResolver) Resolve(
 
 func convert(obj agentv1.Config, branch string) AgentConfiguration {
 	return AgentConfiguration{
-		Assignee:          obj.Spec.Assignee,
-		TaskType:          obj.Spec.TaskType,
-		TaskTypes:         append([]string(nil), obj.Spec.TaskTypes...),
-		Image:             obj.Spec.Image + ":" + branch,
-		Env:               copyEnv(obj.Spec.Env),
-		SecretName:        obj.Spec.SecretName,
-		VolumeClaim:       obj.Spec.VolumeClaim,
-		VolumeMountPath:   obj.Spec.VolumeMountPath,
-		Resources:         obj.Spec.Resources.DeepCopy(),
-		PriorityClassName: obj.Spec.PriorityClassName,
-		Trigger:           obj.Spec.Trigger,
+		Assignee:                obj.Spec.Assignee,
+		TaskType:                obj.Spec.TaskType,
+		TaskTypes:               append([]string(nil), obj.Spec.TaskTypes...),
+		Image:                   obj.Spec.Image + ":" + branch,
+		Env:                     copyEnv(obj.Spec.Env),
+		SecretName:              obj.Spec.SecretName,
+		VolumeClaim:             obj.Spec.VolumeClaim,
+		VolumeMountPath:         obj.Spec.VolumeMountPath,
+		Resources:               obj.Spec.Resources.DeepCopy(),
+		PriorityClassName:       obj.Spec.PriorityClassName,
+		Trigger:                 obj.Spec.Trigger,
+		ZombieJobTimeoutSeconds: obj.Spec.ZombieJobTimeoutSeconds,
 	}
 }